Desafio/requirements.txt
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
protobuf~=4.25.7
python-dateutil~=2.9.0.post0